모임원 0명인 meeting 삭제 (회원 탈퇴, 모임삭제시 모임원 확인)

+++ app/src/main/java/com/example/mmmmeeting/activity/MeetingDeleteActivity.java
@@ -14,6 +14,8 @@ import android.widget.Toast;
 import com.example.mmmmeeting.Info.GridItems;
 import com.example.mmmmeeting.R;
 import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.OnFailureListener;
+import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
@@ -46,7 +48,7 @@ public class MeetingDeleteActivity extends AppCompatActivity {
 
     // 입력한 코드가 존재하는지 확인
     private void checkCode() {
-        final String code = ((EditText)findViewById(R.id.deleteCode)).getText().toString();
+        final String code = ((EditText) findViewById(R.id.deleteCode)).getText().toString();
         DocumentReference docRef = db.collection("meetings").document(code);
 
         docRef.get().addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
@@ -75,43 +77,53 @@ public class MeetingDeleteActivity extends AppCompatActivity {
     private void checkUser(String code) {
         final FirebaseUser user = FirebaseAuth.getInstance().getCurrentUser();
 
-        db.collection("meetings").get()
-                .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
-                    @Override
-                    public void onComplete(@NonNull Task<QuerySnapshot> task) {
-                        if (task.isSuccessful()) {
-                            //모든 document 확인
-                            for (QueryDocumentSnapshot document : task.getResult()) {
-                                // 현재 uid가 모임 정보에 존재하는 경우
-                                if(document.getData().get("userID").toString().contains(user.getUid())){
-                                    // db에서 현재 유저 uid 삭제
-                                    DocumentReference userdel = db.collection("meetings").document(code);
-                                    userdel.update("userID", FieldValue.arrayRemove(user.getUid()));
-                                    startToast("모임에서 탈퇴했습니다.");
-                                    Log.d("Delete", document.getId() + " => " + document.getData());
-                                    myStartActivity(MainActivity.class);
-                                    finish();
-                                    break;
-                                } else {
-                                    Log.d("Delete", "No Document");
-                                    startToast("해당 모임에 가입되어 있지 않습니다.");
-                                    myStartActivity(MainActivity.class);
-                                    finish();
-                                }
+        DocumentReference userdel = db.collection("meetings").document(code);
+
+        userdel.get().addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
+            @Override
+            public void onComplete(@NonNull Task<DocumentSnapshot> task) {
+                if (task.isSuccessful()) {
+                    DocumentSnapshot document = task.getResult();
+                    if (document.exists()) {
+                        // 해당 문서가 존재하는 경우
+                        if (document.getData().get("userID").toString().contains(user.getUid())) {
+                            // db에서 현재 유저 uid 삭제
+                            userdel.update("userID", FieldValue.arrayRemove(user.getUid()));
+                            startToast("모임에서 탈퇴했습니다.");
+                            Log.d("Delete2", document.getId() + " => " + document.getData());
+                            Log.d("Delete2", document.getId() + " => " + document.getData().get("userId"));
+
+                            // 모임원 없는 모임이 된 경우 모임 삭제
+                            if(document.getData().get("userId")==null){
+                                // 모임에 있는 약속, 게시판 내용도 전부 삭제해야..
+                                userdel.delete();
                             }
+
+                            myStartActivity(MainActivity.class);
+                            finish();
                         } else {
-                            Log.d("Delete", "Error getting documents: ", task.getException());
+                            Log.d("Delete", "No Document");
+                            startToast("해당 모임에 가입되어 있지 않습니다.");
+                            myStartActivity(MainActivity.class);
+                            finish();
                         }
+                    } else {
+                        // 존재하지 않는 문서
+                        Log.d("Delete", "No Document");
                     }
-                });
+                } else {
+                    Log.d("Delete", "Task Fail : " + task.getException());
+                }
+            }
+        });
     }
 
     private void startToast(String msg) {
         Toast.makeText(this, msg, Toast.LENGTH_SHORT).show();
     }
 
-    private void myStartActivity(Class c){
-        Intent intent = new Intent(this,c);
+    private void myStartActivity(Class c) {
+        Intent intent = new Intent(this, c);
         intent.addFlags(intent.FLAG_ACTIVITY_CLEAR_TOP);
         startActivity(intent);
     }
